Move bundle uploads to the bundles controller
