+++ lib/cog/models/bundle.ex
@@ -28,7 +28,7 @@ defmodule Cog.Models.Bundle do
     |> Repo.preload(:relay_groups)
     |> cast(params, @required_fields, @optional_fields)
     |> validate_format(:name, ~r/\A[A-Za-z0-9\_\-\.]+\z/)
-    |> unique_constraint(:name, name: :bundles_name_index)
+    |> unique_constraint(:name, name: :bundles_name_index, message: "The bundle name is already in use.")
     |> enable_if_embedded
   end
