+++ mix.exs
@@ -68,7 +68,7 @@ defmodule Cog.Mixfile do
      {:hedwig, "~> 0.3.0"},
      {:gproc, "~> 0.5.0", override: true},
      {:html_entities, "~> 0.3.0"},
-     {:spanner, github: "operable/spanner", ref: "f9ef1de302cd9c497d9269de5e84664cd682d666"},
+     {:spanner, github: "operable/spanner", ref: "2d0771ca370d58a9d76e79003d991c02b014fa83"},
      {:probe, github: "operable/probe", tag: "0.2"},
      {:exml, github: "paulgray/exml", tag: "2.2.1"},
      {:fumanchu, github: "operable/fumanchu"},
