+++ web/router.ex
@@ -36,9 +36,7 @@ defmodule Cog.Router do
     resources "/v1/rules", V1.RuleController, only: [:create, :delete]
 
     resources "/v1/bootstrap", V1.BootstrapController, only: [:index, :create]
-
-    resources "/v1/bundles", V1.BundlesController, only: [:index, :show, :delete]
-    resources "/v1/bundles/deploy", V1.BundleDeployController, only: [:create]
+    resources "/v1/bundles", V1.BundlesController, only: [:index, :show, :delete, :create]
 
     get "/v1/bundles/:id/status", V1.BundleStatusController, :show
     post "/v1/bundles/:id/status", V1.BundleStatusController, :manage_status
